+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Version 3.3.0
 ## [Unreleased]
+#### Added
+- Add PushType.biometric support and BiometricAuthentication class for biometric authentication. Update sample app to handle new Push types [SDKS-1865]
+
 #### Changed
 - Fixed the bug when refreshing the access token we return the old token [SDKS-1824]